+++ R/multi.ace.R
@@ -3,11 +3,9 @@
 #' @description Fast ancestral states estimations run on multiple trees using the Mk model from castor::asr_mk_model.
 #'
 #' @param data A \code{matrix}, \code{data.frame} or \code{list} with the characters for each taxa. Or a \code{multi.ace} list (see details).
-#' @param tree A \code{phylo} or \code{mutiPhylo} object (if the \code{tree} argument contains node labels, they will be used to name the output).
+#' @param tree A \code{phylo} or \code{multiPhylo} object (if the \code{tree} argument contains node labels, they will be used to name the output).
 #' @param models A \code{character} vector, unambiguous named \code{list} or \code{matrix} to be passed as model arguments to \code{castor::asr_mk_model} or \code{ape::ace} (see details).
-#' @param sample An \code{integer} for the number of matrices to sample per tree (default is \code{1}). See details.
-#' @param sample.fun If \code{sample > 1}, a named list containing the following elements: \code{fun} the sampling distribution for continuous characters; and \code{param} (optional) a named list of parameters and their estimation function (default is \code{sample.fun = list(fun = runif, param = list(min = min, max = max))}). See details.
-#' @param threshold Is ignored if \code{sample > 1}, else either a \code{logical} for applying a relative threshold (\code{TRUE} - default) or no threshold (\code{FALSE}) or a \code{numeric} value of the threshold (e.g. 0.95). See details.
+#' @param ml.collapse A named \code{list} for specifying how to collapse maximum likelihood ancestral state estimates. Must contain \code{$type} which can be \code{"majority"} (default; returns the discrete state(s) with highest likelihood, using uncertainty token if tied, and a point estimate (mean) for continuous traits), \code{"relative"} (returns discrete states with scaled likelihood >= max likelihood - 1/n_states), \code{"threshold"} (returns discrete states >= a specified threshold), or \code{"sample"} (samples discrete states by their scaled likelihoods and continuous traits using specified sampling functions). See details for additional elements.
 #' @param special.tokens optional, a named \code{vector} of special tokens to be passed to \code{\link[base]{grep}} (make sure to protect the character with \code{"\\\\"}). By default \code{special.tokens <- c(missing = "\\\\?", inapplicable = "\\\\-", polymorphism = "\\\\&", uncertainty = "\\\\/")}. Note that \code{NA} values are not compared and that the symbol "@" is reserved and cannot be used.
 #' @param special.behaviours optional, a \code{list} of one or more functions for a special behaviour for \code{special.tokens}. See details.
 #' @param brlen.multiplier optional, a vector of branch length modifiers (e.g. to convert time branch length in changes branch length) or a list of vectors (the same length as \code{tree}).
@@ -58,22 +56,16 @@
 #' 
 #' When using the parallel option (either through using \code{parallel = TRUE} by using the number of available cores minus on or manually setting the number of cores - e.g. \code{parallel = 5}), the \code{castor::asr_mk_model} function will use the designated number of cores (using the option \code{Nthreads = <requested_number_of_cores>}). Additionally, if the input \code{tree} is a \code{"multiPhylo"} object, the trees will be run in parallel for each number of cores, thus decreasing computation time accordingly (e.g. if 3 cores are requested and \code{tree} contains 12 \code{"phylo"} objects, 4 different \code{"phylo"} objects will be run in parallel on the 3 cores making the calculation around 3 times faster).
 #'
-#' When using the default \code{sample = 1}, only one estimation is sampled per tree:
-#' \itemize{
-#'      \item For continuous characters, this estimation is the average estimated ancestral value;
-#'      \item For discrete characters, this estimation is the one calculated using the \code{threshold} option (see details below). 
-#' }
-#' When using \code{sample > 1}, multiple estimations are sampled per tree:
-#' \itemize{
-#'      \item For continuous characters, this estimation is sample from the 95% confidence interval using the sampling probability function provided by \code{sample.fun}. By default (\code{runif}), the function samples from a uniform bounded by the 95% confidence interval estimation (see below for modifications).;
-#'      \item For discrete characters, the estimations are sampled using their scaled likelihood.
-#' }
 #'
-#' The \code{threshold} option allows to convert ancestral states likelihoods into discrete states. When \code{threshold = FALSE}, the ancestral state estimated is the one with the highest likelihood (or at random if likelihoods are equal). When \code{threshold = TRUE}, the ancestral state estimated are all the ones that are have a scaled likelihood greater than the maximum observed scaled likelihood minus the inverse number of possible states (i.e. \code{select_state >= (max(likelihood) - 1/n_states)}). This option makes the threshold selection depend on the number of states (i.e. if there are more possible states, a lower scaled likelihood for the best state is expected). Finally using a numerical value for the threshold option (e.g. \code{threshold = 0.95}) will simply select only the ancestral states estimates with a scaled likelihood equal or greater than the designated value. This option makes the threshold selection absolute. Regardless, if more than one value is select, the uncertainty token (\code{special.tokens["uncertainty"]}) will be used to separate the states. If no value is selected, the uncertainty token will be use between all observed characters (\code{special.tokens["uncertainty"]}).
-#'
-#' The \code{sample.fun} option allows to specify a function and parameters for the sampling of the continuous traits. The default is \code{sample.fun = list(fun = runif, param = list(min = min, max = max))} for applying a random uniform sampling (\code{runif}) with the parameters (the minimum and the maximum are applied using respectively the \code{min} and \code{max} functions on the estimated data). For applying different samplings to different traits, you can use a list of arguments in the sample format as \code{sample.fun} (e.g. \code{sample.fun = list(trait_uniform = list(fun = runif, param = list(min = min, max = max)), trait_normal = list(fun = rnorm, param = list(mean = mean, sd = function(x)return(diff(range(x))/4)))} - here the standard deviation is calculated as a quarter of the 95\% CI range).
+#'Additional listed elements for \code{ml.collapse = list()}:
+#' \itemize{
+#'      \item \code{list(type = "majority", tie.breaker = TRUE)} randomly selects one state when highest likelihoods are tied;
+#'      \item \code{list(type = "threshold" , threshold = 0.95)} can take any numeric value (0-1);
+#'      \item \code{list(type = "sample", sample = n, sample.fun = list(fun, param))} where n is an integer > 1 that denotes the number of matrices to sample continuous and/or discrete states across. \code{sample.fun} specifies a function and parameters for the sampling of continuous traits. The default is \code{sample.fun = list(fun = runif, param = list(min = min, max = max))} for applying a random uniform sampling (\code{runif}) with the parameters (the minimum and the maximum are applied using respectively the \code{min} and \code{max} functions on the estimated data). For applying different samplings to different traits, you can use a list of arguments in the sample format as \code{sample.fun} (e.g. \code{sample.fun = list(trait_uniform = list(fun = runif, param = list(min = min, max = max)), trait_normal = list(fun = rnorm, param = list(mean = mean, sd = function(x)return(diff(range(x))/4)))} - here the standard deviation is calculated as a quarter of the 95\% CI range). Note that if \code{type = "sample", sample = 1}, the mean point estimate will be returned if continuous and majority rule applied if discrete.
+#' }
 #' 
-#' It is also possible to just run the estimations without sampling but for sampling later (i.e. just run the estimations; save them as a \code{multi.ace} object and then rerun them into the function). You can do that by using the option \code{output = "multi.ace"}. Using this option ignores the following options: \code{sample}, \code{sample.fun}, and \code{estimation.details}.
+#' 
+#' It is also possible to just run the estimations without sampling but for sampling later (i.e. just run the estimations; save them as a \code{multi.ace} object and then rerun them into the function). You can do that by using the option \code{output = "multi.ace"}. Using this option ignores the following options: \code{ml.collapse}, and \code{estimation.details}.
 #'
 #' @return
 #' Returns a \code{"matrix"} or \code{"list"} of ancestral states. By default, the function returns the ancestral states in the same format as the input \code{matrix}. This can be changed using the option \code{output = "matrix"} or \code{"list"} to force the class of the output.
@@ -84,7 +76,7 @@
 #' If using the option \code{output = "multi.ace"}, the function returns a \code{"multi.ace"} object that can be recycled in the \code{multi.ace} function. This can be useful for testing different sampling strategies of threshold methods without having to recalculate all the ancestral trait estimations.
 #' 
 #' @examples
-#' set.seed(42)
+#' set.seed(34)
 #' ## A simple example:
 #' ## A random tree with 10 tips
 #' tree <- rcoal(10)
@@ -139,7 +131,7 @@
 #' ancestral_states <- multi.ace(matrix_complex, multiple_trees,
 #'                               verbose = TRUE,
 #'                               models = my_models,
-#'                               threshold = 0.95,
+#'                               ml.collapse = list(type = "majority", tie.breaker = TRUE),
 #'                               special.tokens = my_spec_tokens,
 #'                               special.behaviours = my_spec_behaviours,
 #'                               brlen.multiplier = branch_lengths,
@@ -153,7 +145,7 @@
 #' ancestral_states <- multi.ace(matrix_complex, multiple_trees,
 #'                               verbose = TRUE,
 #'                               models = my_models,
-#'                               threshold = 0.95,
+#'                               ml.collapse = list(type = "majority", tie.breaker = TRUE),
 #'                               special.tokens = my_spec_tokens,
 #'                               special.behaviours = my_spec_behaviours,
 #'                               brlen.multiplier = branch_lengths,
@@ -166,11 +158,15 @@
 #'
 #' ## Estimating ancestral states by sampling distributions
 #' ## Discrete characters:
-#' sampled_ancestors <- multi.ace(ancestor_estimates, sample = 25)
+#' sampled_ancestors <- multi.ace(ancestor_estimates,
+#'      ml.collapse = list(type = "sample",
+#'                         sample = 25))
 #'
 #' ## Continuous characters:
 #' ## Sampling 100 matrices from the ancestral estimations
-#' sampled_traits <- multi.ace(matrix_continuous, tree, sample = 100)
+#' sampled_traits <- multi.ace(matrix_continuous, tree,
+#'      ml.collapse = list(type = "sample",
+#'                         sample = 100))
 #' ## Setting up two specific sampling functions
 #' sample_fun <- list(
 #'      unif_sampl = list(fun = runif,
@@ -180,16 +176,18 @@
 #'                                     sd = function(x)return(diff(range(x))/4)))
 #' )
 #' ## Sampling 100 matrices with different distribution functions
-#' sampled_traits <- multi.ace(matrix_continuous, tree, sample = 100,
-#'                             sample.fun = sample_fun)
+#' sampled_traits <- multi.ace(matrix_continuous, tree,
+#'      ml.collapse = list(type = "sample",
+#'                         sample = 100,
+#'                         sample.fun = sample_fun))
 #' 
 #' @seealso
 #' \code{char.diff}, \code{castor::asr_mk_model}, \code{ape::ace}
 #' 
-#' @author Thomas Guillerme
+#' @author Thomas Guillerme, Caleb Scutt
 #' @export
 
-multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = runif, param = list(min = min, max = max)), threshold = TRUE, special.tokens, special.behaviours, brlen.multiplier, verbose = FALSE, parallel = FALSE, output, options.args, estimation.details = NULL) {
+multi.ace <- function(data, tree, models, ml.collapse = list(type = "majority"), special.tokens, special.behaviours, brlen.multiplier, verbose = FALSE, parallel = FALSE, output, options.args, estimation.details = NULL) {
 
     match_call <- match.call()
 
@@ -235,25 +233,55 @@ multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = ru
     
     #########
     ##
-    ## Handle the other options (threshold, brlen, verbose, parallel, output, estimation.details)
+    ## Handle the other options (ml.collapse, brlen, verbose, parallel, output, estimation.details)
     ##
     #########
-
-    ## threshold
-    check.class(threshold, c("logical", "numeric"))
-    if(is(threshold, "logical")) {
-        if(threshold) {
-            ## Use the relative threshold function
-            threshold.type <- "relative"
-        } else {
-            ## Use no threshold (just max)
-            threshold.type <- "max" 
-        }
-    } else {
-        ## Use an absolute threshold
-        threshold.type <- "absolute"
+    do_sample <- FALSE
+    sample <- 1
+    ## ml.collapse
+    check.class(ml.collapse, "list")
+    if(is.null(names(ml.collapse))) {
+        stop(paste0("Invalid ml.collapse option: must be a named list with $type"), call. = FALSE)
     }
+    if(is(ml.collapse, "list")) {
+        valid_types <- c("majority", "relative", "threshold", "sample")
+        if(is.null(ml.collapse$type) || !ml.collapse$type %in% valid_types) {
+        stop(paste0("Invalid ml.collapse$type! Must be one of: ", paste(valid_types, collapse = ", ")), call. = FALSE)
+        }
+
+        if(ml.collapse$type == "majority"){
+            collapse.type <- "majority"
+            if((isTRUE(ml.collapse$tie.breaker))) {
+                collapse.type <- "majority_tiebreaker"
+            }
+        }
+
+        if(ml.collapse$type == "relative"){
+            collapse.type <- "relative"
+        }
     
+        if(ml.collapse$type == "threshold") {
+            ## Use an absolute ml.collapse
+            collapse.type <- "threshold"
+        }
+
+        if(ml.collapse$type == "sample") {
+            ## Validate sample number
+            if(is.null(ml.collapse$sample)) {
+                stop("ml.collapse$sample must be specified when using type = 'sample'", call. = FALSE)
+            }
+            check.class(ml.collapse$sample, c("integer", "numeric"))
+            # check.class(ml.collapse$sample.fun, "list")
+            sample <- ml.collapse$sample
+            do_sample <- sample > 1
+            if(do_sample){
+                collapse.type <- "sample"
+                # sample.fun <- ml.collapse$sample.fun
+            } else {
+                collapse.type <- "majority" ## if sample = 1 reverts to point estimate/majority
+            }
+        }
+    }
     if(!skip_estimations) {
         ## verbose
         check.class(verbose, "logical")
@@ -277,14 +305,14 @@ multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = ru
         }
     }
 
-    ## Sampling
-    check.class(sample, c("integer", "numeric"))
-    do_sample <- sample > 1
-    if(do_sample) {
-        ## Override the threshold arguments (no threshold used)
-        threshold.type <- "sample"
-        threshold <- sample
-    }
+    # # Sampling
+    # check.class(sample, c("integer", "numeric"))
+    # do_sample <- sample > 1
+    # if(do_sample) {
+    #     ## Override the ml.collapse arguments (no ml.collapse used)
+    #     collapse.type <- "sample"
+    #     ml.collapse <- sample
+    # }
     
     if(!skip_estimations) {
         ## brlen multiplier
@@ -867,6 +895,11 @@ multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = ru
         } else {
             ## Check the sampling (if required)
             if(do_sample) {
+                if(is.null(ml.collapse$sample.fun)) {
+                    sample.fun <- list(fun = runif, param = list(min = min, max = max))
+                } else {
+                    sample.fun <- ml.collapse$sample.fun
+                }
                 sample.fun_class <- check.class(sample.fun, "list")
                 ## If sample.fun is a single list
                 if(!is.null(names(sample.fun)) && names(sample.fun)[1] == "fun") {
@@ -917,38 +950,53 @@ multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = ru
         ancestral_estimations <- lapply(discrete_estimates, lapply, function(estimation) return(estimation$results))
 
         ## Apply the selector
-        switch(threshold.type,
-            relative = {select.states <- function(taxon, threshold) {
+        switch(collapse.type,
+            relative = {select.states <- function(taxon, ml.collapse) {
                                                   if(all(is.na(taxon))) {
                                                     return(NA)
                                                   } else {
                                                     return(names(taxon[taxon >= (max(taxon) - 1/length(taxon))]))
                                                   }
                                                  }},
-            max      = {select.states <- function(taxon, threshold) {
+            majority = {select.states <- function(taxon, ml.collapse) {
+                                  if(all(is.na(taxon))) {
+                                    return(NA)
+                                  } else {
+                                    return(names(taxon[taxon >= (max(taxon))]))
+                                  }
+                             }},
+
+            majority_tiebreaker = {select.states <- function(taxon, ml.collapse) {
                                                   if(all(is.na(taxon))) {
                                                     return(NA)
                                                   } else {
-                                                    return(names(taxon[taxon >= (max(taxon))]))
+                                                    max_states <- names(taxon[taxon >= (max(taxon))])
+                                                    if (length(max_states) > 1) {
+                                                        return(sample(max_states, size = 1)) ## randomly selects one state if tied (always returns single states)
+                                                    } else { 
+                                                        return(max_states)
+                                                    }
                                                   }
                                                  }},
-            absolute = {select.states <- function(taxon, threshold) {
+            threshold = {select.states <- function(taxon, ml.collapse) {
                                                   if(all(is.na(taxon))) {
                                                     return(NA)
                                                   } else {
+                                                    threshold <- ml.collapse$threshold
                                                     return(names(taxon[taxon >= threshold]))
                                                   }
                                                  }},
-            sample  = {select.states <- function(taxon, threshold){
+            sample  = {select.states <- function(taxon, ml.collapse){
                                                  if(all(is.na(taxon))) {
                                                     return(NA)
                                                  } else {
-                                                    return(sample(names(taxon), size = threshold, prob = taxon, replace = TRUE))
+                                                    n <- ml.collapse$sample
+                                                    return(sample(names(taxon), size = n, prob = taxon, replace = TRUE))
                                                  }}}
                                              )
 
         ## Get the ancestral states
-        ancestral_states <- lapply(ancestral_estimations, lapply, translate.likelihood, threshold, select.states, special.tokens, do_sample)
+        ancestral_states <- lapply(ancestral_estimations, lapply, translate.likelihood, ml.collapse, select.states, special.tokens, do_sample) ## needs fixing
 
         ## Add the invariant characters
         if(has_invariants) {
@@ -1074,4 +1122,5 @@ multi.ace <- function(data, tree, models, sample = 1, sample.fun = list(fun = ru
 # serial_time <- serial_end-serial_start
 # parallel_time <- parallel_end-parallel_start
 # claddis_time <- claddis_end-claddis_start
+## Set seed for reproducibility